Boeken kunnen aan bieb toegevoegd worden, verwijderd worden, en worden laten zien

+++ bieb-app-productie-code/bieb-app/src/main/java/com/leesenleer/BiebApp/services/ReadCommunityService.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ReadCommunityService extends CrudRepository<ReadCommunity, Integer> {
+public interface ReadCommunityService extends CrudRepository<ReadCommunity, Long> {
 }
